requirements.txt
streamlit
pandas
numpy
Pillow
torch
torchvision
opencv-python-headless
ultralytics